fixing conversion warning on "tests" directory

+++ tests/xmltester/XMLTester.cpp
@@ -127,7 +127,7 @@ double java_math_round(double val)
 const char * getIndent( unsigned int numIndents )
 {
     static const char * pINDENT = "                                      + ";
-    static const unsigned int LENGTH = strlen( pINDENT );
+    static const unsigned int LENGTH = static_cast<unsigned int>(strlen( pINDENT ));
 
     if ( numIndents > LENGTH ) numIndents = LENGTH;
